requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
sqlmodel==0.0.8
psycopg2-binary==2.9.11

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-dotenv==1.0.0

algosdk==2.7.0
httpx==0.27.2
python-multipart==0.0.20
requests